# Use LinkedList in scope locals