@cthulahoops 2019-08-31
If it's doubly-linked doesn't this break tail sharing? (Ie. the clone requires a copy of the list?)

This code is still the wrong structure, so I don't think we should be making this kind of small optimisation until we've fixed the big problems.