@cthulahoops 2019-08-31
To be more precise load_global should be able to skip over the list of scopes and just access the globals. The call to last should go away.