@palaviv 2019-08-31
I think I should have stated that one of the main goals of this change is to use `use std::collections::LinkedList` instead of implementing one our self. I think this simplifies the code and make it more easy to understand.

 > If it's doubly-linked doesn't this break tail sharing? (Ie. the clone requires a copy of the list?)

You are correct but as can be seen from the performance benchmarks this does not have a negative effect. We have a list of referenced counted object creating a copy of the list is very cheap.

> To be more precise load_global should be able to skip over the list of scopes and just access the globals. The call to last should go away.

Even without the fast access to last I think the simplification to the code is worth the change.