@cthulahoops 2019-08-31
> I think I should have stated that one of the main goals of this change is to use `use std::collections::LinkedList` instead of implementing one our self. I think this simplifies the code and make it more easy to understand.

It's nice to see the code reduction, but I'm not sure it's worth using the wrong data structure for. I'm guessing there's no standard library implementation of RcList?

> You are correct but as can be seen from the performance benchmarks this does not have a negative effect. We have a list of referenced counted object creating a copy of the list is very cheap.

The benchmarks don't use nested scopes at all. mandelbrot doesn't even have a local scope. That said, very deeply nesting scopes isn't common in python, so the list should usually be short. (The slow down would be on calling closures, I think.)