@windelbouwman 2019-09-01
I propose to merge this after #1333 gets merged. I like the usage of a standard linked list implementation, to reduce code.