@windelbouwman 2019-09-01
Too bad we do no longer share the tail of the list, but I think this is okay for now, and can be added back later on, if required.