@cthulahoops 2019-09-02
Trying out a more relevant benchmark suggests that this branch isn't significantly slower. (Confirming the suggestion above that the short linked list copy is cheap.)

Master:
test bench_rustpy_closure       ... bench:      38,520 ns/iter (+/- 2,478)

This branch:
test bench_rustpy_closure       ... bench:      38,823 ns/iter (+/- 713)

I still worry that the copy here will get forgotten / need undoing, especially as the list elements may not be simple reference counted objects once local scopes are implemented properly.