@coolreader18 2019-09-03
What about just a plain `Vec`? If we're not doing tail-sharing, I don't think a `LinkedList` has all that many benefits compared to a `Vec`.