@palaviv 2019-09-04
> What about just a plain `Vec`? If we're not doing tail-sharing, I don't think a `LinkedList` has all that many benefits compared to a `Vec`.

Now that we do not access the last value `LinkedList` is not needed. I changed it to `Vec`